refactor: 상품 목록 위젯 분리
