+++ src/refactoring/widget/product-list/index.ts
@@ -1,0 +1,1 @@
+export { ProductList } from "./ui";
